initial refactor work
