--- maya-stubgen/maya_stubgen/cli/generate_stubs/parsers/__init__.py
@@ -1,3 +1,0 @@
-from .builtin_parser import *
-from .cmds_parsers import *
-from .common import *
