+++ maya-stubgen/maya_stubgen/parsers/__init__.py
@@ -1,0 +1,4 @@
+from .builtin_parser import *
+from .maya.cmds import *
+from .common import *
+from .parser import *
