+++ maya-stubgen/pyproject.toml
@@ -16,6 +16,7 @@ docspec-to-jinja = { path = "../docspec-to-jinja/", develop = true }
 lxml = "^4.9.1"
 requests = "^2.28.1"
 typing-extensions = "^4.8.0"
+httpx = "^0.27.0"
 
 [tool.poetry.dev-dependencies]
 snakeviz = "^2.1.1"
